ecom_item_sv/requirements.txt
Django
djangorestframework
mongoengine
psycopg2
gunicorn
requests